requirements.txt
flask==2.3.3
tensorflow-cpu==2.13.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.0.0
gunicorn==21.2.0